Support configuring link buffer sizes.
This adds support for configuring link buffer sizes in the virtual
topology. Buffer size is configured in number of packets, and only
active when bandwidth limitations are in effect. If no buffer size is
configured, a reasonable default is calculated based on the configured
BDP if delay is set, or defaults to 1000 packets otherwise.

This is only implemented for the netns backend. As part of this change,
TBF is no longer used for shaping links; it is replaced by netem's
'rate' parameter.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>